refactor(cli)!: remove `vectorcode hooks` in favour of `vectorcode init --hooks`
