+++ src/vectorcode/subcommands/__init__.py
@@ -2,7 +2,6 @@ from vectorcode.subcommands.check import check
 from vectorcode.subcommands.chunks import chunks
 from vectorcode.subcommands.clean import clean
 from vectorcode.subcommands.drop import drop
-from vectorcode.subcommands.hooks import hooks
 from vectorcode.subcommands.init import init
 from vectorcode.subcommands.ls import ls
 from vectorcode.subcommands.prompt import prompts
@@ -15,7 +14,6 @@ __all__ = [
     "chunks",
     "clean",
     "drop",
-    "hooks",
     "init",
     "ls",
     "prompts",
